StringOfItem constructor from string
