+++ commonItems/BlobList.cs
@@ -1,0 +1,43 @@
+﻿using System.Collections.Generic;
+using System.Text;
+
+namespace commonItems;
+
+public class BlobList : Parser {
+	public List<string> Blobs { get; } = new();
+	public BlobList(BufferedReader reader) {
+		var next = GetNextLexeme(reader);
+		if (next == "=") {
+			next = GetNextLexeme(reader);
+		}
+		if (next != "{") {
+			return;
+		}
+
+		var braceDepth = 0;
+		var sb = new StringBuilder();
+		while (!reader.EndOfStream) {
+			char inputChar = (char)reader.Read();
+			if (inputChar == '{') {
+				if (braceDepth > 0) {
+					sb.Append(inputChar);
+				}
+				++braceDepth;
+			} else if (inputChar == '}') {
+				--braceDepth;
+				if (braceDepth > 0) {
+					sb.Append(inputChar);
+				} else if (braceDepth == 0) {
+					Blobs.Add(sb.ToString());
+					sb.Clear();
+				} else if (braceDepth == -1) {
+					return;
+				}
+			} else if (braceDepth == 0) {
+				// Ignore this character. Only look for blobs.
+			} else {
+				sb.Append(inputChar);
+			}
+		}
+	}
+}
